Date picker, fixed modal

+++ resources/assets/js/components/Collaborateurs.vue
@@ -4,7 +4,7 @@
         <div class="row pt-5">
 
             <div class="col-lg-12 d-flex justify-content-end pb-2">
-                <button type="button" class="btn btn-success" data-toggle="modal" data-target="#create-item">
+                <button type="button" @click="ajouterBtn()" class="btn btn-success" data-toggle="modal" data-target="#create-item" >
                     Ajouter collaborateur <i class="fa fa-plus" aria-hidden="true"></i>
                 </button>
             </div>
@@ -15,13 +15,12 @@
                 <div class="card mb-3">
                     <img class="card-img-top" :src="'http://localhost:8000/'+collaborateur.image" alt="Card image cap">
                     <div class="card-body">
-                        <h4 class="card-title">{{ collaborateur.nom }}</h4>
-                        <p class="card-text">{{ collaborateur.prenom }}</p>
-                        <p class="card-text">{{ collaborateur.email }}</p>
-                        <p class="card-text">{{ collaborateur.telephone }}</p>
-                        <p class="card-text">{{ collaborateur.post }}</p>
-                        <p class="card-text">{{ collaborateur.adresse }}</p>
-                        <p class="card-text">{{ collaborateur.date_naissance }}</p>
+                        <h4 class="card-title">{{ collaborateur.nom }} {{collaborateur.prenom}}</h4>
+                        <p class="card-text"> <strong> Email </strong> : {{ collaborateur.email }}</p>
+                        <p class="card-text"> <strong> Tel </strong>: {{ collaborateur.telephone }}</p>
+                        <p class="card-text"> <strong> Post </strong> : {{ collaborateur.post }}</p>
+                        <p class="card-text"> <strong> Adresse </strong>: {{ collaborateur.adresse }}</p>
+                        <p class="card-text"> <strong> Date de naissance </strong>: {{ collaborateur.date_naissance }}</p>
 
 
                         <button type="button" class="btn btn-primary" @click.prevent="editCollaborateur( collaborateur )" ><i class="fa fa-pencil" aria-hidden="true"></i></button>
@@ -31,7 +30,7 @@
             </div>
 
             <!-- Create Collab Modal -->
-            <div class="modal fade" id="create-item" tabindex="-1" role="dialog" aria-labelledby="exampleModalLabel" aria-hidden="true">
+            <div class="modal fade" id="create-item" tabindex="-1" role="dialog" aria-labelledby="exampleModalLabel" aria-hidden="true" data-backdrop="static" data-keyboard="false">
                 <div class="modal-dialog" role="document">
                     <div class="modal-content">
                         <div class="modal-header">
@@ -42,10 +41,12 @@
                         </div>
                         <form method="post" enctype="multipart/form-data" @submit.prevent="createCollaborateur">
                             <div class="modal-body">
+
+
                                 <div class="col-lg-12">
                                     <div class="form-group">
                                         <label>Image</label>
-                                        <input type="file" @change="imageChangedOnCreate" :src="'http://localhost:8000/'+newItem.image"  class="form-control">
+                                        <input type="file" data-preview="#preview" @change="imageChangedOnCreate" :src="'http://localhost:8000/'+newItem.image"  class="form-control">
                                     </div>
                                 </div>
                                 <div class="col-lg-12">
@@ -74,20 +75,24 @@
                                 </div>
                                 <div class="col-lg-12">
                                     <div class="form-group">
-                                        <label>Email</label>
-                                        <input type="text" class="form-control" v-model="newItem.email">
+                                        <label >Email <span class="required">*</span> </label>
+                                       <input name= "email" v-model="newItem.email" required  v-validate="'required|email'" :class="{'input': true, 'is-danger': errors.has('email') }" class="form-control" >
+                                        <span v-show="errors.has('email')" style="color: red" class="text-danger">{{ errors.first('email') }}</span>
+
                                     </div>
                                 </div>
                                 <div class="col-lg-12">
                                     <div class="form-group">
                                         <label>Telephone</label>
-                                        <input type="text" class="form-control" v-model="newItem.telephone">
+                                        <input name="telephone" v-validate="'required|numeric'" :class="{'input': true, 'is-danger': errors.has('telephone') }" type="text" class="form-control" v-model="newItem.telephone">
+                                        <i v-show="errors.has('telephone')" class="fa fa-warning"></i>
+                                        <span v-show="errors.has('telephone')" class="" style="color: red">{{ errors.first('telephone') }}</span>
                                     </div>
                                 </div>
                                 <div class="col-lg-12">
                                     <div class="form-group">
                                         <label>Date de naissance</label>
-                                        <input type="text" class="form-control" v-model="newItem.date_naissance">
+                                        <flat-pickr :config="config" placeholder="Choisissez une date"  class="form-control" v-model="newItem.date_naissance"> </flat-pickr>
                                     </div>
                                 </div>
                             </div>
@@ -101,7 +106,7 @@
             </div>
 
             <!-- Edit Collab Modal -->
-            <div class="modal fade" id="edit-item" tabindex="-1" role="dialog" aria-labelledby="exampleModalLabel" aria-hidden="true">
+            <div class="modal fade" id="edit-item" tabindex="-1" role="dialog" aria-labelledby="exampleModalLabel" aria-hidden="true" data-backdrop="static" data-keyboard="false">
                 <div class="modal-dialog" role="document">
                     <div class="modal-content">
                         <div class="modal-header">
@@ -176,7 +181,8 @@
 </template>
 
 <script>
-
+    import flatPickr from 'vue-flatpickr-component';
+    import 'flatpickr/dist/flatpickr.css';
     export default {
 
         data(){
@@ -184,10 +190,18 @@
             return {
 
                 newItem: { 'nom': '', 'prenom': '', 'image': '', 'post': '', 'date_naissance': '', 'email':'', 'telephone': '', 'adresse': '', 'image': '' },
-                 fillItem: { 'nom': '', 'prenom': '', 'image': '', 'post': '', 'date_naissance': '', 'email':'', 'telephone': '', 'adresse': '' , 'image': '','id': ''},
-                collaborateurs: []
+                fillItem: { 'nom': '', 'prenom': '', 'image': '', 'post': '', 'date_naissance': '', 'email':'', 'telephone': '', 'adresse': '' , 'image': '','id': ''},
+                collaborateurs: [],
+                date: new Date(),
+                config: {
+                    dateFormat: 'Y-m-d',
+
+                },
 
             }
+        },
+        components: {
+            flatPickr
         },
 
         mounted() {
@@ -209,15 +223,17 @@
                     this.collaborateurs.push(response.data);
                     this.newItem = { 'nom': '', 'prenom': '', 'image': '', 'post': '', 'date_naissance': '', 'email':'', 'telephone': '', 'adresse': '' };
                     $('#create-item').modal('hide');
+                    this.getCollaborateurs();
 
                 }).catch( (error) => {
                     //console.log( error.response.data )
                 })
             },
-            editCollaborateur( collaborateur ) {
+            editCollaborateur(collaborateur) {
                 let edit = this.fillItem;
 
                 edit.nom = collaborateur.nom;
+                edit.id = collaborateur.id;
                 edit.prenom = collaborateur.prenom;
                 edit.adresse = collaborateur.adresse;
                 edit.post = collaborateur.post;
@@ -231,7 +247,7 @@
             updateCollaborateur(id){
                 let input = this.fillItem;
 
-                axios.put('vue-collaborateurs/'+id, input).then( (response)=> {
+                axios.put('vue-collaborateurs/' + id, input).then( (response)=> {
 
                     this.getCollaborateurs();
                     this.fillItem = { 'nom': '', 'prenom': '', 'image': '', 'post': '', 'date_naissance': '', 'email':'', 'telephone': '', 'adresse': '' , 'id': ''};
@@ -258,12 +274,28 @@
                     this.fillItem.image = e.target.result
                 }
             },
-
+            ajouterBtn(){
+                this.newItem = { 'nom': '', 'prenom': '', 'image': '', 'post': '', 'date_naissance': '', 'email':'', 'telephone': '', 'adresse': '' };
+            },
             deleteCollaborateur( collaborateur ) {
-                axios.delete('vue-collaborateurs/' + collaborateur.id).then( (response) => {
-                    this.getCollaborateurs();
-                });
+                let conf = confirm("Do you ready want to delete this collab?");
+                if (conf === true) {
+                    axios.delete('vue-collaborateurs/' + collaborateur.id).then((response) => {
+                        this.getCollaborateurs();
+                    });
+                }
             }
         }
     }
 </script>
+
+<style>
+    .card-img-top {
+        width: 100%;
+        height: 15vw;
+        object-fit: cover;
+    }
+    .required {
+        color: red;
+    }
+</style>
